+++ CHANGES.md
@@ -3,6 +3,7 @@ Markdown Writer FX Change Log
 
 ## 0.8
 
+- Open markdown files by dropping them to main window.
 - Support Definition lists extension (FlexMark only).
 - "Cut", "Copy", "Paste" and "Select All" commands added to "Edit" menu.
 - "Cut" and "Copy" now cut/copy whole line if selection is empty.